NS record working
